requirements.txt
flask>=3.0.0
twilio>=9.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0
pytz>=2024.1
gunicorn>=21.0.0
solders>=0.18.0
base58>=2.1.1
web3>=6.0.0
